Conflicts resolved by keeping setup-and-maintenance-update improvements (type
annotations, numpy-style docstrings, logging, ruff compliance) and integrating
the functional changes from main:

- src/crane_controller/envs/controlled_crane_pendulum.py: apply dtype=np.float64
  fix to continuous observation space (PR #3)
- src/crane_controller/ppo_agent.py: integrate VecNormalize wrapping, fix dead-code
  bug in __init__ (inference_only flag), add _stats_path helper, save vecnorm stats
  in do_training, normalize obs in do_one_episode (PR #3)
- scripts/train_ppo.py: add VecNormalize stats path log after training (PR #3)
- tests/test_environment.py: add test_observation_space_dtype and
  test_observations_are_float (PR #3)
- tests/test_ppo.py: add test_ppo_saves_vecnorm and test_ppo_vecnorm_updates (PR #3)
- tests/conftest.py: change fixture scope from "session" to "package" for chdir
  and test_dir (PR #4)
- tests/test_algorithm.py: accept auto-merged improvements from PR #4
- 8 stale test data files removed: anti-pendulum*.json, pendulum*.json,
  antipendulum.zip, ppo_AntiPendulumEnv.zip, interval.json (PR #4)

+++ src/crane_controller/envs/controlled_crane_pendulum.py
@@ -144,7 +144,7 @@ class AntiPendulumEnv(gym.Env[AntiPendulumObs, int]):
             self.discrete = {}
             self.spaces_min = np.array((-size, -max_speed, 0.0, -max_speed), float)
             self.spaces_max = np.array((size, max_speed, np.pi, max_speed), float)
-            self.observation_space = spaces.Box(self.spaces_min, self.spaces_max, shape=(4,), dtype=np.int64)
+            self.observation_space = spaces.Box(self.spaces_min, self.spaces_max, shape=(4,), dtype=np.float64)
 
         self.nresets: int = 0
         self.acc = acc
